backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
google-generativeai==0.3.2
openai==1.3.7
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
chromadb==0.4.18
langchain==0.0.350
langchain-openai==0.0.2
opencv-python==4.8.1.78
asyncio-mqtt==0.16.1
